Renamed package constant to util
