+++ src/test/java/mate/academy/service/impl/BookServiceImplTest.java
@@ -1,24 +1,24 @@
 package mate.academy.service.impl;
 
-import static mate.academy.constant.TestConstants.BOOK_NOT_FOUND_MESSAGE;
-import static mate.academy.constant.TestConstants.CATEGORY_NOT_FOUND_MESSAGE;
-import static mate.academy.constant.TestConstants.DUPLICATE_ISBN_MESSAGE;
-import static mate.academy.constant.TestConstants.FIRST_RECORD;
-import static mate.academy.constant.TestConstants.SAMPLE_BOOK_ID;
-import static mate.academy.constant.TestUtil.BOOK_DTO_WITHOUT_CATEGORY_IDS;
-import static mate.academy.constant.TestUtil.BOOK_PAGE;
-import static mate.academy.constant.TestUtil.CATEGORY_DTO;
-import static mate.academy.constant.TestUtil.CREATE_BOOK_REQUEST_DTO;
-import static mate.academy.constant.TestUtil.FIRST_CATEGORY;
-import static mate.academy.constant.TestUtil.NEW_BOOK;
-import static mate.academy.constant.TestUtil.PAGEABLE;
-import static mate.academy.constant.TestUtil.SAMPLE_BOOK;
-import static mate.academy.constant.TestUtil.SAMPLE_BOOK_DTO;
-import static mate.academy.constant.TestUtil.SEARCH_PARAMETERS;
-import static mate.academy.constant.TestUtil.SECOND_CATEGORY;
-import static mate.academy.constant.TestUtil.UPDATED_BOOK;
-import static mate.academy.constant.TestUtil.UPDATED_BOOK_DTO;
-import static mate.academy.constant.TestUtil.UPDATE_BOOK_REQUEST_DTO;
+import static mate.academy.util.TestConstants.BOOK_NOT_FOUND_MESSAGE;
+import static mate.academy.util.TestConstants.CATEGORY_NOT_FOUND_MESSAGE;
+import static mate.academy.util.TestConstants.DUPLICATE_ISBN_MESSAGE;
+import static mate.academy.util.TestConstants.FIRST_RECORD;
+import static mate.academy.util.TestConstants.SAMPLE_BOOK_ID;
+import static mate.academy.util.TestUtil.BOOK_DTO_WITHOUT_CATEGORY_IDS;
+import static mate.academy.util.TestUtil.BOOK_PAGE;
+import static mate.academy.util.TestUtil.CATEGORY_DTO;
+import static mate.academy.util.TestUtil.CREATE_BOOK_REQUEST_DTO;
+import static mate.academy.util.TestUtil.FIRST_CATEGORY;
+import static mate.academy.util.TestUtil.NEW_BOOK;
+import static mate.academy.util.TestUtil.PAGEABLE;
+import static mate.academy.util.TestUtil.SAMPLE_BOOK;
+import static mate.academy.util.TestUtil.SAMPLE_BOOK_DTO;
+import static mate.academy.util.TestUtil.SEARCH_PARAMETERS;
+import static mate.academy.util.TestUtil.SECOND_CATEGORY;
+import static mate.academy.util.TestUtil.UPDATED_BOOK;
+import static mate.academy.util.TestUtil.UPDATED_BOOK_DTO;
+import static mate.academy.util.TestUtil.UPDATE_BOOK_REQUEST_DTO;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.atLeastOnce;
