+++ src/test/java/mate/academy/util/TestUtil.java
@@ -1,41 +1,41 @@
-package mate.academy.constant;
+package mate.academy.util;
 
-import static mate.academy.constant.TestConstants.CATEGORY_DESCRIPTION;
-import static mate.academy.constant.TestConstants.CATEGORY_ID;
-import static mate.academy.constant.TestConstants.CATEGORY_NAME;
-import static mate.academy.constant.TestConstants.NEW_BOOK_AUTHOR;
-import static mate.academy.constant.TestConstants.NEW_BOOK_CATEGORY_IDS;
-import static mate.academy.constant.TestConstants.NEW_BOOK_COVER_IMAGE;
-import static mate.academy.constant.TestConstants.NEW_BOOK_DESCRIPTION;
-import static mate.academy.constant.TestConstants.NEW_BOOK_ISBN;
-import static mate.academy.constant.TestConstants.NEW_BOOK_PRICE;
-import static mate.academy.constant.TestConstants.NEW_BOOK_TITLE;
-import static mate.academy.constant.TestConstants.NEW_CATEGORY_DESCRIPTION;
-import static mate.academy.constant.TestConstants.NEW_CATEGORY_NAME;
-import static mate.academy.constant.TestConstants.SAMPLE_BOOK_AUTHOR;
-import static mate.academy.constant.TestConstants.SAMPLE_BOOK_COVER_IMAGE;
-import static mate.academy.constant.TestConstants.SAMPLE_BOOK_DESCRIPTION;
-import static mate.academy.constant.TestConstants.SAMPLE_BOOK_ID;
-import static mate.academy.constant.TestConstants.SAMPLE_BOOK_ISBN;
-import static mate.academy.constant.TestConstants.SAMPLE_BOOK_PRICE;
-import static mate.academy.constant.TestConstants.SAMPLE_BOOK_TITLE;
-import static mate.academy.constant.TestConstants.SEARCH_AUTHOR;
-import static mate.academy.constant.TestConstants.SEARCH_FIELD_AUTHOR;
-import static mate.academy.constant.TestConstants.SEARCH_PAGE_NUMBER;
-import static mate.academy.constant.TestConstants.SEARCH_PAGE_SIZE;
-import static mate.academy.constant.TestConstants.SEARCH_PRICE_MAX;
-import static mate.academy.constant.TestConstants.SEARCH_PRICE_MIN;
-import static mate.academy.constant.TestConstants.SEARCH_TITLE;
-import static mate.academy.constant.TestConstants.SECOND_CATEGORY_DESCRIPTION;
-import static mate.academy.constant.TestConstants.SECOND_CATEGORY_ID;
-import static mate.academy.constant.TestConstants.SECOND_CATEGORY_NAME;
-import static mate.academy.constant.TestConstants.UPDATED_BOOK_AUTHOR;
-import static mate.academy.constant.TestConstants.UPDATED_BOOK_CATEGORY_IDS;
-import static mate.academy.constant.TestConstants.UPDATED_BOOK_COVER_IMAGE;
-import static mate.academy.constant.TestConstants.UPDATED_BOOK_DESCRIPTION;
-import static mate.academy.constant.TestConstants.UPDATED_BOOK_ISBN;
-import static mate.academy.constant.TestConstants.UPDATED_BOOK_PRICE;
-import static mate.academy.constant.TestConstants.UPDATED_BOOK_TITLE;
+import static mate.academy.util.TestConstants.CATEGORY_DESCRIPTION;
+import static mate.academy.util.TestConstants.CATEGORY_ID;
+import static mate.academy.util.TestConstants.CATEGORY_NAME;
+import static mate.academy.util.TestConstants.NEW_BOOK_AUTHOR;
+import static mate.academy.util.TestConstants.NEW_BOOK_CATEGORY_IDS;
+import static mate.academy.util.TestConstants.NEW_BOOK_COVER_IMAGE;
+import static mate.academy.util.TestConstants.NEW_BOOK_DESCRIPTION;
+import static mate.academy.util.TestConstants.NEW_BOOK_ISBN;
+import static mate.academy.util.TestConstants.NEW_BOOK_PRICE;
+import static mate.academy.util.TestConstants.NEW_BOOK_TITLE;
+import static mate.academy.util.TestConstants.NEW_CATEGORY_DESCRIPTION;
+import static mate.academy.util.TestConstants.NEW_CATEGORY_NAME;
+import static mate.academy.util.TestConstants.SAMPLE_BOOK_AUTHOR;
+import static mate.academy.util.TestConstants.SAMPLE_BOOK_COVER_IMAGE;
+import static mate.academy.util.TestConstants.SAMPLE_BOOK_DESCRIPTION;
+import static mate.academy.util.TestConstants.SAMPLE_BOOK_ID;
+import static mate.academy.util.TestConstants.SAMPLE_BOOK_ISBN;
+import static mate.academy.util.TestConstants.SAMPLE_BOOK_PRICE;
+import static mate.academy.util.TestConstants.SAMPLE_BOOK_TITLE;
+import static mate.academy.util.TestConstants.SEARCH_AUTHOR;
+import static mate.academy.util.TestConstants.SEARCH_FIELD_AUTHOR;
+import static mate.academy.util.TestConstants.SEARCH_PAGE_NUMBER;
+import static mate.academy.util.TestConstants.SEARCH_PAGE_SIZE;
+import static mate.academy.util.TestConstants.SEARCH_PRICE_MAX;
+import static mate.academy.util.TestConstants.SEARCH_PRICE_MIN;
+import static mate.academy.util.TestConstants.SEARCH_TITLE;
+import static mate.academy.util.TestConstants.SECOND_CATEGORY_DESCRIPTION;
+import static mate.academy.util.TestConstants.SECOND_CATEGORY_ID;
+import static mate.academy.util.TestConstants.SECOND_CATEGORY_NAME;
+import static mate.academy.util.TestConstants.UPDATED_BOOK_AUTHOR;
+import static mate.academy.util.TestConstants.UPDATED_BOOK_CATEGORY_IDS;
+import static mate.academy.util.TestConstants.UPDATED_BOOK_COVER_IMAGE;
+import static mate.academy.util.TestConstants.UPDATED_BOOK_DESCRIPTION;
+import static mate.academy.util.TestConstants.UPDATED_BOOK_ISBN;
+import static mate.academy.util.TestConstants.UPDATED_BOOK_PRICE;
+import static mate.academy.util.TestConstants.UPDATED_BOOK_TITLE;
 
 import java.util.List;
 import java.util.Set;
